requirements.txt
streamlit==1.29.0
PyMuPDF==1.23.7
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
groq==0.4.1
streamlit-lottie==0.0.5
python-dotenv==1.0.0
httpx>=0.24.0
python-dotenv